Update: basic data types to unsigned int

+++ BigInt.h
@@ -10,8 +10,7 @@ typedef unsigned char byte;
 class BigInt {
 public:
     BigInt();
-    // BigInt(byte a);
-    BigInt(int a);
+    BigInt(unsigned int a);
 	~BigInt() {}
 	friend BigInt operator+(const BigInt& lop, const BigInt& rop);// reference: https://stackoverflow.com/questions/269268/how-to-implement-big-int-in-c
 	friend BigInt operator-(const BigInt& lop, const BigInt& rop);
@@ -27,12 +26,12 @@ public:
 	friend bool operator!= (const BigInt& lop, const BigInt& rop);
 	std::string toHex();// reference: https://stackoverflow.com/questions/5661101/how-to-convert-an-unsigned-character-array-into-a-hexadecimal-string-in-c
 	int getLength() const;
-	void put(byte value); // put value to the rear
-	void insertZeros(int num); // insert zero to the front
-	void insert(byte value); // insert value to high pos
-	byte get(int index) const;
+	void put(unsigned int value); // put value to the rear
+	void insertZeros(int num); // insert zeros to the front
+	void insert(unsigned int value); // insert value to high pos
+	unsigned int get(int index) const;
 	void setValue(std::string value);
-	void setValue(byte value); // set single byte value
+	void setValue(unsigned int value); // set single byte value
 
 private:
 	void trim();
@@ -46,6 +45,4 @@ const BigInt TWO_BIG_INT = BigInt(2);
 BigInt pow(BigInt base, BigInt power);
 BigInt pow(BigInt base, BigInt power, BigInt modulo); // reference: https://www.khanacademy.org/computing/computer-science/cryptography/modarithmetic/a/fast-modular-exponentiation
 BigInt getRandom(BigInt lowerbound, BigInt upperbound); // both inclusive
-
-// BigInt trimmed();
 #endif
